Add files via upload
The file contains the entire script of the source code for the Arduino/ESP control board and the script necessary for the website to work (in HTML format). The program is used to control a prototype dog food feeder, all in order to feed your quadruped without moving from your place. Being in a big hurry and not having enough time, you can use WiFi and your computer or smartphone to control the doors in the feeder. This is a great convenience and time saver, especially for people who are very busy or have several dogs.

The aim of the project was not only to create a prototype feeder, but a building that works in practice with doors moving in the vertical axis. The whole thing is controlled wirelessly using a smartphone or computer. Feedback for the user about statistics and the current position of the door is included. In addition, this information is to be signaled by LED diodes. The feeder consists of 3 floors with food and one for the microprocessor system.