RTOS project
The aim of the project was to understand the operation of a real-time system. The task was to write a program that would allow for setting a time-varying load on the system. Writing it was based on the previous program, which used the basic API functions. Additional load tasks were created. Each of these tasks had a delay loop, which caused the system to overload. However, not every task received information about the set load in the same way. Five of them did it using a queue, five more using a box, and the remaining five by capturing a semaphore. The program was forbidden to use casting, so each assignment had to be done in a thoughtful way. All the most important information about the current operation of the program was displayed on the screen, in real time.