Visual Information Processing - project
One of the most common tasks in image processing is changing the color of an image. This often means removing a dominant or unwanted color cast. The aim of the project was to implement a color correction algorithm based on Reinhart's method, which involves transferring colors from one image to another. To achieve this, the color space had to be converted from RGB, where the colors are highly correlated, to the lαβ space, which minimizes this correlation. After converting to the lαβ space and performing several transformations, the implemented method allows for transferring the color scheme from one image to another.

To implement the Reinhart method, the Matlab environment was used. After loading the test images, it was necessary to isolate the appropriate R, G, B values for each pixel, and then create a matrix consisting of these three values for each pixel. In the following steps, various transformations were performed to convert the color space from RGB, where the colors are highly correlated, to the lαβ space, which minimizes this correlation. In this space, color correction could be performed. For this purpose, the means and standard deviations along each axis were determined for both images. Then, for each pixel of the base image, its mean was subtracted, and the resulting value was multiplied by the ratio of the standard deviations of the source image to the base image. Finally, the previously subtracted mean was added back, but from the source image rather than the base image. By reversing the previous transformations, the base image returns to the RGB color space, but with the altered colors, and all three images are displayed: the base image, the source image, and the newly transformed image.