Learning C Language - University Studies
The files contain multiple scripts written in the C language, created by me during my studies in the course "Computer Programming for Computations". The number of folders is 10 because there were 10 lessons in this course. The aim of the course was to provide students with practical methods for problem analysis and algorithmization of solutions, necessary for independently developing programs that perform engineering calculations on standard PC computers using the C language.