Learning databases - University Studies
The objective of the project classes is to consolidate the knowledge presented in the lecture from the previous semester. During the execution of various tasks, students have the opportunity to practice SQL queries (both simple and complex) on a database specifically prepared for the project classes, design a database schema for a given fragment of reality described as a potential client would, issue appropriate DDL commands to create necessary database objects, and learn about mechanisms unique to "classical" relational systems functioning in PostgreSQL, an example of a so-called post-relational system. Additionally, students learn how to write an application in a high-level language (Java) that will interact with the chosen database.