Digital image processing - University Studies
The aim of the course is to familiarize students with the basic methods of digital image processing and recognition. Students will learn techniques for processing visual signals, and after attending the lecture series, they will create their own computer applications based on the acquired knowledge and using source literature, as part of laboratory exercises.