Embedded systems - University Studies
The aim of the course is to provide practical and theoretical knowledge about the principles of microprocessor systems and the possibilities of using these systems in modern control systems. Additionally, the course covers the design, construction, and implementation of microprocessor systems, as well as the development of embedded and system software. During laboratory exercises, students have the opportunity to gain practical knowledge using the most popular 8-bit and 32-bit microcontrollers.